# server/docs/filehandler-openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "File Sharing API",
    "description" : "API for File Sharing Application - Upload, Download, Share files",
    "contact" : {
      "name" : "File Sharing Team",
      "email" : "support@filesharing.com"
    },
    "license" : {
      "name" : "MIT License",
      "url" : "https://opensource.org/licenses/MIT"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://localhost:5000",
    "description" : "Local Development Server"
  }, {
    "url" : "https://api.filesharing.com",
    "description" : "Production Server"
  } ],
  "security" : [ {
    "Bearer Authentication" : [ ]
  } ],
  "paths" : {
...
    "/api/review-session/{reviewSessionId}/decision" : {
      "post" : {
        "tags" : [ "review-session-controller" ],
        "operationId" : "submitDecision",
        "parameters" : [ {
          "name" : "reviewSessionId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReviewSessionDecisionDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonResponseReviewSessionEntity"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized - Invalid or missing JWT token"
          },
          "403" : {
            "description" : "Forbidden - Access denied"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
    "/api/review-session/create" : {
      "post" : {
        "tags" : [ "review-session-controller" ],
        "operationId" : "createReviewSession",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReviewSessionCreateDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonResponseReviewSessionEntity"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized - Invalid or missing JWT token"
          },
          "403" : {
            "description" : "Forbidden - Access denied"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
...
    "/api/review-session/asset/{assetId}" : {
      "get" : {
        "tags" : [ "review-session-controller" ],
        "operationId" : "getReviewSessionsByAsset",
        "parameters" : [ {
          "name" : "assetId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "versionNumber",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonResponseListReviewSessionEntity"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized - Invalid or missing JWT token"
          },
          "403" : {
            "description" : "Forbidden - Access denied"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ReviewSessionDecisionDTO" : {
        "type" : "object",
        "properties" : {
          "decision" : {
            "$ref" : "#/components/schemas/ReviewSessionStatus"
          },
          "note" : {
            "type" : "string"
          }
        }
      },
      "ReviewSessionStatus" : {
        "type" : "string",
        "enum" : [ "DRAFT", "IN_REVIEW", "REQUEST_CHANGES", "APPROVED" ]
      },
      "CommonResponseReviewSessionEntity" : {
        "type" : "object",
        "properties" : {
          "isSuccessful" : {
            "type" : "boolean"
          },
          "data" : {
            "$ref" : "#/components/schemas/ReviewSessionEntity"
          },
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        }
      },
      "ReviewMetrics" : {
        "type" : "object",
        "properties" : {
          "totalAnnotations" : {
            "type" : "integer",
            "format" : "int32"
          },
          "openAnnotations" : {
            "type" : "integer",
            "format" : "int32"
          },
          "resolvedAnnotations" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalComments" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "ReviewSessionEntity" : {
        "type" : "object",
        "properties" : {
          "isTrash" : {
            "type" : "boolean"
          },
          "trashedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "createdBy" : {
            "type" : "string"
          },
          "createdByEmail" : {
            "type" : "string"
          },
          "updateBy" : {
            "type" : "string"
          },
          "updateByEmail" : {
            "type" : "string"
          },
          "isActive" : {
            "type" : "boolean"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "reviewSessionId" : {
            "type" : "string"
          },
          "projectId" : {
            "type" : "string"
          },
          "assetId" : {
            "type" : "string"
          },
          "versionNumber" : {
            "type" : "integer",
            "format" : "int32"
          },
          "title" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "dueDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "status" : {
            "$ref" : "#/components/schemas/ReviewSessionStatus"
          },
          "statusHistory" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ReviewStatusHistory"
            }
          },
          "reviewers" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ReviewerInfo"
            }
          },
          "metrics" : {
            "$ref" : "#/components/schemas/ReviewMetrics"
          },
          "completedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "ReviewStatusHistory" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "$ref" : "#/components/schemas/ReviewSessionStatus"
          },
          "changedBy" : {
            "type" : "string"
          },
          "changedByEmail" : {
            "type" : "string"
          },
          "changedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "note" : {
            "type" : "string"
          }
        }
      },
      "ReviewerInfo" : {
        "type" : "object",
        "properties" : {
          "userId" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "role" : {
            "$ref" : "#/components/schemas/ReviewerRole"
          },
          "invitedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "lastViewedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "hasCommented" : {
            "type" : "boolean"
          }
        }
      },
      "ReviewerRole" : {
        "type" : "string",
        "enum" : [ "REVIEWER", "APPROVER" ]
      },
      "ReviewSessionCreateDTO" : {
        "type" : "object",
        "properties" : {
          "projectId" : {
            "type" : "string"
          },
          "assetId" : {
            "type" : "string"
          },
          "versionNumber" : {
            "type" : "integer",
            "format" : "int32"
          },
          "title" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "dueDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "reviewerIds" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
...
      "CommonResponseListReviewSessionEntity" : {
        "type" : "object",
        "properties" : {
          "isSuccessful" : {
            "type" : "boolean"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ReviewSessionEntity"
            }
          },
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "Bearer Authentication" : {
        "type" : "http",
        "description" : "Enter JWT token",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}